+++ src/components/button/index.tsx
@@ -27,7 +27,7 @@ Props) {
       className={
         disabled
           ? "w-full flex justify-center bg-gray-300 p-3 px-6 rounded-md text-white"
-          : "w-full flex justify-center  bg-primary_green-500 text-white p-3 px-6 rounded-md hover:bg-transparent hover:border hover:border-primary_green-500 transition-all hover:text-primary_green-500"
+          : "w-full flex justify-center  bg-primary text-white p-3 px-6 rounded-md hover:bg-transparent hover:border hover:border-primary transition-all hover:text-primary"
       }
       // onClick={() => handleClick()}
     >