+++ src/components/logo/index.tsx
@@ -1,0 +1,8 @@
+export default function Logo() {
+  return (
+    <div className=" flex items-center gap-3 text-primary">
+      <img src="/logo_blue.png" alt="logo" className="" />
+      <h5 className=" font-extrabold text-xl md:text-3xl">99Apartment</h5>
+    </div>
+  );
+}