+++ tailwind.config.js
@@ -3,6 +3,7 @@ module.exports = {
   content: [
     "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
+    "./src/assets/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
     "./src/layouts/**/*.{js,ts,jsx,tsx,mdx}",
   ],
@@ -15,7 +16,7 @@ module.exports = {
       },
       colors: {
         primary: {
-          default: "#2E4393",
+          DEFAULT: "#2E4393",
         },
       },
     },